requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
opencv-python>=4.8.0
torchvision>=0.15.0
pyyaml>=6.0
pytest>=7.0.0
